Change QNT value type from Integer to Double in the database and in the apps code.

+++ src/models/DepotStock.java
@@ -2,21 +2,20 @@ package models;
 
 public class DepotStock extends Depot{
 
-	private int qnt;
+	private double qnt;
 
-	public DepotStock(Depot depot, int qnt){
+	public DepotStock(Depot depot, double qnt){
 		super(depot.getCode(), depot.getName(), depot.getComments());
 		this.qnt = qnt;
 	}
 
 	
-	public int getQnt() {
+	public double getQnt() {
 		return qnt;
 	}
 
-	public void setQnt(int qnt) {
+	public void setQnt(double qnt) {
 		this.qnt = qnt;
 	}
-
 
 }
